Login screen created and signUp updated
In this commit we have created Login screen that has basic functionality of sign In In the app and showing user name in the home screen. We updated the sign Up screen to create username for user. As of this point app has only basic ui. This will be updated in the end of project.