Refactor Styling and Enhance Movie List Search Behavior
In this commit, we have created a separate stylesheet for reusable styles that are shared across multiple components. This includes the styles for the SignUp and Login components, as well as the styles used in the FlatList within the Home and Watchlist components. Additionally, we improved the user experience by modifying the logic for the movie list. Now, when the user scrolls down the list and performs a search, the movie list will reset and display from the beginning, which was not the case before.